+++ packages/bridge/src/messaging/index.ts
@@ -1,0 +1,2 @@
+export * from "./messaging.types";
+export * from "./messaging";