+++ packages/swirl-tokens/dart/lib/styles.dark.dart
@@ -4,7 +4,7 @@
 //
 
 // Do not edit directly
-// Generated on Thu, 24 Nov 2022 12:39:43 GMT
+// Generated on Wed, 14 Dec 2022 09:43:39 GMT